+++ Unity-Package/Assets/root/Editor/Scripts/Data/GetParticleSystemResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2fadae63d851a0449634c5b27666009
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
