+++ Unity-Package/Assets/root/Editor/Scripts/Tools/ParticleSystem.Get.cs
@@ -18,7 +18,7 @@ using com.IvanMurzak.Unity.MCP.Runtime.Extensions;
 using com.IvanMurzak.Unity.MCP.Utils;
 using Microsoft.Extensions.Logging;
 
-namespace com.IvanMurzak.Unity.MCP.ParticleSystem.Editor.API
+namespace com.IvanMurzak.Unity.MCP.ParticleSystem.Editor
 {
     public partial class Tool_ParticleSystem
     {
@@ -121,6 +121,12 @@ namespace com.IvanMurzak.Unity.MCP.ParticleSystem.Editor.API
             bool deepSerialization = false
         )
         {
+            if (gameObjectRef == null)
+                throw new ArgumentNullException(nameof(gameObjectRef));
+
+            if (!gameObjectRef.IsValid(out var gameObjectValidationError))
+                throw new ArgumentException(gameObjectValidationError, nameof(gameObjectRef));
+
             return MainThread.Instance.Run(() =>
             {
                 var go = gameObjectRef.FindGameObject(out var error);
@@ -128,7 +134,7 @@ namespace com.IvanMurzak.Unity.MCP.ParticleSystem.Editor.API
                     throw new Exception(error);
 
                 if (go == null)
-                    throw new Exception(Error.GameObjectNotFound());
+                    throw new Exception("GameObject not found.");
 
                 // Find the ParticleSystem component
                 UnityEngine.ParticleSystem? ps = null;
@@ -141,7 +147,7 @@ namespace com.IvanMurzak.Unity.MCP.ParticleSystem.Editor.API
                     if (comp == null)
                         continue;
 
-                    if (componentRef != null && componentRef.IsValid)
+                    if (componentRef != null && componentRef.IsValid(out _))
                     {
                         if (componentRef.Matches(allComponents[i], i))
                         {
@@ -160,7 +166,7 @@ namespace com.IvanMurzak.Unity.MCP.ParticleSystem.Editor.API
                 }
 
                 if (ps == null)
-                    throw new Exception(Error.ParticleSystemNotFound());
+                    throw new Exception("ParticleSystem component not found on the specified GameObject.");
 
                 var response = new GetParticleSystemResponse
                 {
